pyproject.toml
[project]
name = "immich-e-ink-frame"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.4",
    "pillow>=11.1.0",
    "pillow-heif>=0.22.0",
    "pip>=25.0.1",
    "python-daemon>=3.1.2",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "spotipy>=2.25.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]
